Refactor runtime error handling and improve source text management
- Updated error handling in web UI to provide more detailed runtime errors.
- Changed the error type in the runtime compiler to use `RuntimeWireResult` for better consistency.
- Enhanced the `SourceText` structure to use interned strings (`SourceAtom`) for improved memory efficiency.
- Added caching for full text retrieval in `SourceText` to optimize performance.
- Refactored various functions to accommodate changes in error handling and source text management.
- Introduced new methods for span validation and added utility functions for better code clarity.
- Improved tests to cover new functionality and ensure robustness.

+++ src/grammar/display.rs
@@ -253,10 +253,10 @@ pub fn format_grammar_error_message(error: &GrammarError) -> String {
 fn format_error_context(
     out: &mut String,
     lines: &[&str],
-    start_line: usize,
-    end_line: usize,
-    start_col: usize,
-    end_col: usize,
+    start_line: u32,
+    end_line: u32,
+    start_col: u32,
+    end_col: u32,
 ) {
     use std::fmt::Write;
 
@@ -266,10 +266,14 @@ fn format_error_context(
 
     let line_idx = start_line - 1; // Convert to 0-indexed
     let context_before = if line_idx > 0 { 1 } else { 0 };
-    let context_after = if line_idx + 1 < lines.len() { 1 } else { 0 };
+    let context_after = if line_idx + 1 < lines.len() as u32 {
+        1
+    } else {
+        0
+    };
 
     let first_line = line_idx.saturating_sub(context_before);
-    let last_line = (line_idx + context_after).min(lines.len() - 1);
+    let last_line = (line_idx + context_after).min(lines.len() as u32 - 1);
     let gutter_width = format!("{}", last_line + 1).len().max(1);
 
     // Show lines before
@@ -278,7 +282,7 @@ fn format_error_context(
             out,
             "\n{:>width$} | {}",
             i + 1,
-            lines[i],
+            lines[i as usize],
             width = gutter_width
         );
     }
@@ -289,7 +293,7 @@ fn format_error_context(
         "\n{}{:>width$} | {}{}",
         RED,
         line_idx + 1,
-        lines[line_idx],
+        lines[line_idx as usize],
         RESET,
         width = gutter_width
     );
@@ -297,13 +301,13 @@ fn format_error_context(
     // Show underline/arrow
     let underline_indent = gutter_width + 3; // gutter + " | "
     let _ = write!(out, "\n{}", " ".repeat(underline_indent));
-    let _ = write!(out, "{}", " ".repeat(start_col));
+    let _ = write!(out, "{}", " ".repeat(start_col as usize));
 
     // Show underline spanning the error
     if start_line == end_line && start_col < end_col {
         let len = end_col - start_col;
         let _ = write!(out, "{}", RED);
-        let _ = write!(out, "{}", "~".repeat(len.max(1)));
+        let _ = write!(out, "{}", "~".repeat(len.max(1) as usize));
         let _ = write!(out, "{}", RESET);
     } else {
         let _ = write!(out, "{}^{}", RED, RESET);
@@ -315,7 +319,7 @@ fn format_error_context(
             out,
             "\n{:>width$} | {}",
             i + 1,
-            lines[i],
+            lines[i as usize],
             width = gutter_width
         );
     }
